refactor, tests run again, now need to make them pass
