+++ package.json
@@ -50,6 +50,7 @@
     "chai": "^4.1.2",
     "dirty-chai": "^2.0.1",
     "gulp": "^3.9.1",
+    "libp2p-websocket-star-rendezvous": "^0.1.2",
     "lodash": "^4.17.4"
   },
   "repository": {
