+++ test/clean.js
@@ -1,4 +1,12 @@
+/* eslint-env mocha */
+
 let toclean = []
+
+afterEach(() => {
+  toclean.forEach(s => s.disconnect())
+  toclean = []
+})
+
 function cleaner () {
   toclean.forEach(s => {
     if (s.disconnect) s.disconnect()
@@ -6,11 +14,8 @@ function cleaner () {
   })
 }
 
-afterEach(() => {
-  toclean.forEach(s => s.disconnect())
-  toclean = []
-})
-module.exports = function () {
+module.exports = () => {
   toclean.push.apply(toclean, arguments)
 }
+
 module.exports.cleaner = cleaner
