+++ test/instance.spec.js
@@ -6,15 +6,15 @@ const dirtyChai = require('dirty-chai')
 const expect = chai.expect
 chai.use(dirtyChai)
 
-const WSStar = require('../../src')
+const WebSocketStar = require('../src')
 
 describe('instantiate the transport', () => {
   it('create', () => {
-    const wstar = new WSStar()
+    const wstar = new WebSocketStar()
     expect(wstar).to.exist()
   })
 
   it('create without new', () => {
-    expect(() => WSStar()).to.throw()
+    expect(() => WebSocketStar()).to.throw()
   })
 })
