+++ test/node.js
@@ -1,20 +1,3 @@
-/* eslint-env mocha */
 'use strict'
 
-const WStar = require('..')
-
-require('./sig-server.js')
-
-describe('transport', () => {
-  const create = () => {
-    return new WStar({})
-  }
-
-  require('./transport/dial.js')(create)
-  require('./transport/listen.js')(create)
-  require('./transport/discovery.js')(create)
-  require('./transport/filter.js')(create)
-  require('./transport/strict.js')(create)
-  require('./transport/valid-connection.js')(create)
-  require('./transport/reconnect.node.js')(create)
-})
+require('./reconnect.node.js')
